+++ src/widgets/box.ts
@@ -1,25 +1,24 @@
 import AgsWidget, { type BaseProps } from './widget.js';
-import GObject from 'gi://GObject';
 import Gtk from 'gi://Gtk?version=3.0';
-import Service from '../service.js';
 
-export interface BoxProps<T> extends BaseProps<T>, Gtk.Box.ConstructorProperties {
+export type BoxProps = BaseProps<AgsBox, Gtk.Box.ConstructorProperties & {
     children?: Gtk.Widget[]
     vertical?: boolean
-}
+}>;
 
 export default class AgsBox extends AgsWidget(Gtk.Box) {
     static {
-        GObject.registerClass({
-            GTypeName: 'AgsBox',
-            Properties: {
-                'vertical': Service.pspec('vertical', 'boolean', 'rw'),
-                'children': Service.pspec('children', 'jsobject', 'rw'),
+        AgsWidget.register(this, {
+            properties: {
+                'vertical': ['boolean', 'rw'],
+                'children': ['jsobject', 'rw'],
             },
-        }, this);
+        });
     }
 
-    constructor(props: BoxProps<AgsBox> = {}) { super(props); }
+    constructor(props: BoxProps = {}) {
+        super(props as Gtk.Box.ConstructorProperties);
+    }
 
     get children() { return this.get_children(); }
     set children(children: Gtk.Widget[]) {